test/requirements.txt
bidict==0.23.1
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
msgpack-python==0.5.6
PyJWT==2.10.1
python-engineio==4.13.0
python-socketio==5.16.0
requests==2.32.5
simple-websocket==1.1.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
python-decouple==3.8
